WEB-4460 move filters into directory
