+++ __tests__/unit/app/pages/clinicworkspace/TideDashboardV2/filters/FilterBySites.test.js
@@ -6,7 +6,7 @@ import configureStore from 'redux-mock-store';
 import { thunk } from 'redux-thunk';
 import { MemoryRouter } from 'react-router-dom';
 
-import FilterBySites from '@app/pages/clinicworkspace/TideDashboardV2/FilterBySites';
+import FilterBySites from '@app/pages/clinicworkspace/TideDashboardV2/filters/FilterBySites';
 
 const mockStore = configureStore([thunk]);
