+++ app/pages/clinicworkspace/TideDashboardV2/TideDashboardV2.js
@@ -4,10 +4,12 @@ import { useSelector } from 'react-redux';
 import Table from '../../../components/elements/Table';
 import { Flex, Text, Box } from 'theme-ui';
 
-import FilterByCategory from './FilterByCategory';
-import FilterByTags from './FilterByTags';
-import FilterByDataRecency from './FilterByDataRecency';
-import FilterBySummaryPeriod from './FilterBySummaryPeriod';
+import FilterByCategory from './filters/FilterByCategory';
+import FilterByTags from './filters/FilterByTags';
+import FilterBySites from './filters/FilterBySites';
+import FilterByDataRecency from './filters/FilterByDataRecency';
+import FilterBySummaryPeriod from './filters/FilterBySummaryPeriod';
+import AppliedFiltersList from './filters/AppliedFiltersList';
 
 import TableCategoryHeader from './TableCategoryHeader';
 import PaginationController from './PaginationController';
@@ -18,8 +20,6 @@ import useTableColumns from './useTableColumns';
 import EmptyContentNode from './EmptyContentNode';
 import { Redirect, useLocation } from 'react-router-dom';
 import useAuthorizationGate from './useAuthorizationGate';
-import FilterBySites from './FilterBySites';
-import AppliedFiltersList from './AppliedFiltersList';
 
 const Gap = () => <Box sx={{ marginLeft: 'auto' }}></Box>;
