+++ app/pages/clinicworkspace/TideDashboardV2/filters/FilterByCategory.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector, useDispatch } from 'react-redux';
-import { SegmentedControl, Segment } from '../components/SegmentedControl';
-import { setCategory, setOffset, CATEGORY } from './tideDashboardSlice';
+import { SegmentedControl, Segment } from '../../components/SegmentedControl';
+import { setCategory, setOffset, CATEGORY } from '../tideDashboardSlice';
 import { colors as vizColors } from '@tidepool/viz';
 import { Box } from 'theme-ui';
